requirements.txt
mlflow==3.6.0
scikit-learn==1.3.2
pandas==2.1.3
numpy>=1.26.0
pydantic==2.5.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
pylint==3.0.3
requests==2.31.0
prometheus-client==0.19.0
psutil==6.0.0
psycopg2-binary==2.9.9

